requirements.txt
cycler==0.10.0
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.0
numpy==1.16.3
Pillow==6.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
